Uncompleted drone riding mechanic for code

+++ kervangamesp1/Assets/!Scripts/Enemy/Drone/Drone.cs
@@ -27,7 +27,7 @@ public class Drone : Enemy
         ChangeState(awakeState);
     }
 
-    void TakeDamage(float damage){
+    void TakeDamage(int damage){
         
     }
